+++ docs/src/examples/components/Button/Types/ButtonExampleEmphasis.shorthand.tsx
@@ -3,8 +3,8 @@ import { Button } from '@stardust-ui/react'
 
 const ButtonExampleEmphasis = () => (
   <div>
-    <Button type="primary" content="Primary" />
-    <Button type="secondary" content="Secondary" />
+    <Button content="Primary" primary />
+    <Button content="Secondary" secondary />
   </div>
 )