+++ docs/src/examples/components/Button/Variations/ButtonExampleEmphasisCircular.tsx
@@ -3,10 +3,10 @@ import { Button, Icon } from '@stardust-ui/react'
 
 const ButtonExampleEmphasisCircular = () => (
   <div>
-    <Button type="primary" circular icon>
+    <Button circular icon primary>
       <Icon name="coffee" xSpacing="none" />
     </Button>
-    <Button type="secondary" circular icon>
+    <Button circular icon secondary>
       <Icon name="book" xSpacing="none" />
     </Button>
   </div>